+++ src/components/renderer/OneBlinkFormElements.tsx
@@ -109,7 +109,12 @@ function OneBlinkFormElements<T extends FormTypes._NestedElementsElement>({
           return (
             <div
               key={element.id}
-              className="ob-element cypress-element-container"
+              className={clsx(
+                'ob-element cypress-element-container',
+                element.customCssClasses
+                  ? element.customCssClasses.join(' ')
+                  : '',
+              )}
             >
               <FormElementSection
                 formId={formId}